+++ src/storage/AnimCore.ts
@@ -363,7 +363,9 @@ export let AnimCore = class AnimCore {
 			trigger: this.CONST.TRIGGERS[0],
 			preset: this.CONST.PRESETS[0],
 			file: '',
-			options: {},
+			options: {
+				tieToDocuments: true,
+			},
 		}),
 		PRESETS: Object.keys(presets) as PresetKeys[],
 		TRIGGERS: [