+++ src/view/_components/AnimationEditor.svelte
@@ -1,89 +1,80 @@
 <script lang='ts'>
+	import type { TJSDocument } from '@typhonjs-fvtt/runtime/svelte/store/fvtt/document'
+	import derivedFlag from 'src/lib/docFlagDerived'
 	import { AnimCore, type JSONData } from 'src/storage/AnimCore'
-	import { writable } from 'svelte/store'
-	import { onDestroy } from 'svelte'
-	import SubAnimationEditor from './SubAnimationEditor.svelte'
+	import { type Writable, writable } from 'svelte/store'
+	import { ErrorMsg } from 'src/utils'
 	import JSONEditorApp from './JSONEditor/JSONEditor'
+	import SingleEditor from './SingleEditor.svelte'
 
-	export let key: string
-	export let editable: boolean = true
-	export let value: Exclude<JSONData[string], string>
-	export let deleteFn: false | ((key: string) => void) = false
+	export let doc: TJSDocument<ActorPF2e | ItemPF2e | UserPF2e> | Writable<{ id: 'settings' }>
+	export function createAnimation() {}
 
-	async function deleteSubAnimation(animation: Exclude<JSONData[string], string>[number]) {
-		Dialog.confirm({
-			content: `Are you sure you want to delete this animation?`,
-			title: 'Confirm Deletion',
-			yes: () => {
-				value = value.filter(x => x !== animation)
-			},
-		})
+	const flag = $doc.id === 'settings'
+		? window.pf2eGraphics.storeSettings.getWritableStore('worldAnimations') as Writable<JSONData>
+		// @ts-ignore Above is a type guard
+		: derivedFlag(doc, 'pf2e-graphics', 'customAnimations', {} as JSONData, 1000)
+
+	if (!flag) throw new ErrorMsg('No document was provided to AnimationEditor?!')
+
+	let search = ''
+	let newAnimeKey = ''
+	const inStasis = writable(false)
+
+	function openJSON() {
+		new JSONEditorApp({
+			id: `pf2e-graphics-${$doc.id}`,
+			store: flag,
+			stasis: inStasis,
+			permission: 'canUserModify' in $doc ? $doc.canUserModify(game.user, 'update') : true,
+		}).render(true)
 	}
 
-	async function deleteAnimation() {
-		Dialog.confirm({
-			content: `Are you sure you want to delete all animations under "${key}"?`,
-			title: 'Confirm Deletion',
-			yes: () => {
-				if (deleteFn) {
-					deleteFn(key)
-				} else {
-					// @ts-ignore Explicitly a deletion thing
-					value = null
-				}
-			},
-		})
+	function addNew() {
+		if (newAnimeKey.trim().length) {
+			$flag[newAnimeKey.trim()] ??= [AnimCore.CONST.TEMPLATE_ANIMATION()]
+			newAnimeKey = ''
+		} else if ('getRollOptions' in $doc) {
+			const prefix = $doc.isOfType('weapon') ? 'item' : undefined
+			newAnimeKey = (($doc as ItemPF2e).getRollOptions(prefix))[2]
+		} else {
+			newAnimeKey = ' '
+		}
 	}
-
-	const store = writable(value)
-	store.subscribe((v) => {
-		value = v
-	})
-
-	$: store.set(value)
-
-	let app: null | JSONEditorApp = null
-	async function exportWindow() {
-		app = new JSONEditorApp(
-			{ data: { store, key } },
-		).render(true, {
-			focus: true,
-		})
-	}
-
-	onDestroy(() => {
-		app?.close({ force: true })
-	})
 </script>
 
-<div class='border border-solid bg-slate-300 bg-opacity-50 w-full rounded-sm p-2.5'>
-	<label class='flex items-center gap-1'>
-		<span>Primary Roll Option:</span>
-		<input type='text' value={key} disabled class='w-min text-center' />
+<div class='
+	flex
+	border border-solid rounded-sm
+	mb-1 p-1
+	overflow-clip
+'>
+	<div class='[&>*:not(:first-child)]:mx-1 my-auto flex items-center text-nowrap'>
+		<span><i class='fa fa-search' /> Search</span><input bind:value={search} />
+	</div>
 
-		<div class='ml-auto flex items-center'>
-			{#if editable}
-				<button
-					on:click={() => {
-						value = [...value, AnimCore.CONST.TEMPLATE_ANIMATION()]
-					}}
-					class='w-min text-nowrap items-center'
-				>
-					<i class='fa fa-plus' />
-					New Child Animation
-				</button>
-			{/if}
-			<button data-tooltip='pf2e-graphics.export' on:click={exportWindow} class='fa fa-brackets-curly size-8' />
-			<button data-tooltip='Delete' on:click={deleteAnimation} class='fa fa-trash-can size-8' />
-		</div>
-	</label>
-
-	{#if value.length}
-		<hr />
-		<div class='flex flex-col gap-1.5'>
-			{#each value as animation}
-				<SubAnimationEditor bind:animation {deleteSubAnimation} />
-			{/each}
-		</div>
+	<button
+		class='fa fa-brackets-curly size-8 ml-auto'
+		data-tooltip='pf2e-graphics.jsonEditor'
+		on:click={openJSON}
+	/>
+	{#if newAnimeKey.length}
+		<input bind:value={newAnimeKey} />
 	{/if}
+	<button
+		class='fa fa-add size-8 text-lg leading-4'
+		data-tooltip='pf2e-graphics.addAnim'
+		on:click={addNew}
+		disabled={$inStasis}
+	/>
+</div>
+<div class='flex flex-col gap-1'>
+	{#each Object.keys($flag).filter(k => k.includes(search)) as key}
+		<SingleEditor
+			bind:key
+			bind:value={$flag[key]}
+			bind:disabled={$inStasis}
+			flag={flag}
+		/>
+	{/each}
 </div>